+++ packages/core/fixtures/app/env.d.ts
