+++ packages/core/src/adapters/server-utils.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from 'bun:test';
-import { FIXTURE_PROJECT_DIR } from '../../fixtures/constants';
-import { AppConfigurator } from '../main/app-configurator';
-import { ServerUtils } from './server-utils.module';
+import { FIXTURE_PROJECT_DIR } from '../../fixtures/constants.ts';
+import { AppConfigurator } from '../main/app-configurator.ts';
+import { ServerUtils } from './server-utils.module.ts';
 
 await AppConfigurator.create({
   projectDir: FIXTURE_PROJECT_DIR,
