+++ packages/core/src/main/app-configurator.ts
@@ -1,10 +1,10 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import { appLogger } from '../global/app-logger';
-import type { EcoPagesAppConfig } from '../internal-types';
-import type { EcoPagesConfig } from '../public-types';
-import { deepMerge } from '../utils/deep-merge';
-import { invariant } from '../utils/invariant';
+import { appLogger } from '../global/app-logger.ts';
+import type { EcoPagesAppConfig } from '../internal-types.ts';
+import type { EcoPagesConfig } from '../public-types.ts';
+import { deepMerge } from '../utils/deep-merge.ts';
+import { invariant } from '../utils/invariant.ts';
 
 export class AppConfigurator {
   config: EcoPagesAppConfig;
