+++ packages/core/src/main/build-app.ts
@@ -1,11 +1,11 @@
 import { PostCssProcessor } from '@ecopages/postcss-processor';
-import { appLogger } from '../global/app-logger';
-import { AppBuilder } from '../main/app-builder';
-import { AppConfigurator } from '../main/app-configurator';
-import { CssBuilder } from '../main/css-builder';
-import { IntegrationManager } from '../main/integration-manager';
-import { ScriptsBuilder } from '../main/scripts-builder';
-import { StaticPageGenerator } from '../main/static-page-generator';
+import { appLogger } from '../global/app-logger.ts';
+import { AppBuilder } from '../main/app-builder.ts';
+import { AppConfigurator } from '../main/app-configurator.ts';
+import { CssBuilder } from '../main/css-builder.ts';
+import { IntegrationManager } from '../main/integration-manager.ts';
+import { ScriptsBuilder } from '../main/scripts-builder.ts';
+import { StaticPageGenerator } from '../main/static-page-generator.ts';
 
 export async function buildApp({
   config = process.cwd(),
