+++ packages/core/tsconfig.json
@@ -1,6 +1,8 @@
 {
 	"compilerOptions": {
 		"allowSyntheticDefaultImports": true,
+		"allowImportingTsExtensions": true,
+		"noEmit": true,
 		"baseUrl": ".",
 		"rootDir": ".",
 		"esModuleInterop": true,
